+++ TaskMenu/Utilities/Constants.swift
@@ -48,6 +48,7 @@ enum Constants {
         static let automaticUpdateChecksEnabledKey = "automaticUpdateChecksEnabled"
         static let lastUpdateCheckDateKey = "lastUpdateCheckDate"
         static let lastAlertedUpdateVersionKey = "lastAlertedUpdateVersion"
+        static let menuBarTitleListIdKey = "menuBarTitleListId"
     }
 
     enum Notifications {